move investigationView state under tab
