+++ src/js/state/View/actions.js
@@ -1,12 +1,6 @@
 /* @flow */
 
-import type {
-  DOWNLOADS_HIDE,
-  DOWNLOADS_SHOW,
-  INVESTIGATION_VIEW_SET,
-  InvestigationView,
-  TIME_ZONE_SET
-} from "./types"
+import type {DOWNLOADS_HIDE, DOWNLOADS_SHOW, TIME_ZONE_SET} from "./types"
 
 export default {
   setTimeZone: (timeZone: string): TIME_ZONE_SET => ({
@@ -20,10 +14,5 @@ export default {
 
   hideDownloads: (): DOWNLOADS_HIDE => ({
     type: "DOWNLOADS_HIDE"
-  }),
-
-  setInvestigationView: (view: InvestigationView): INVESTIGATION_VIEW_SET => ({
-    type: "INVESTIGATION_VIEW_SET",
-    view
   })
 }
