+++ src/js/state/View/types.js
@@ -1,17 +1,11 @@
 /* @flow */
 
-export type InvestigationView = "tree" | "linear"
 export type ViewState = {
   downloadsIsOpen: boolean,
-  timeZone: string,
-  investigationView: InvestigationView
+  timeZone: string
 }
 
-export type ViewAction =
-  | TIME_ZONE_SET
-  | DOWNLOADS_SHOW
-  | DOWNLOADS_HIDE
-  | INVESTIGATION_VIEW_SET
+export type ViewAction = TIME_ZONE_SET | DOWNLOADS_SHOW | DOWNLOADS_HIDE
 
 export type TIME_ZONE_SET = {
   type: "TIME_ZONE_SET",
@@ -24,9 +18,4 @@ export type DOWNLOADS_SHOW = {
 
 export type DOWNLOADS_HIDE = {
   type: "DOWNLOADS_HIDE"
-}
-
-export type INVESTIGATION_VIEW_SET = {
-  type: "INVESTIGATION_VIEW_SET",
-  view: InvestigationView
 }
